+++ src/Ext/Techno/Hooks.Misc.cpp
@@ -431,4 +431,37 @@ DEFINE_HOOK(0x51D7E0, InfantryClass_DoAction_Water, 0x5)
 	return Continue;
 }
 
+DEFINE_HOOK(0x730D1F, DeployCommandClass_Execute_VoiceDeploy, 0x5)
+{
+	GET_STACK(int, unitsToDeploy, STACK_OFFSET(0x18, -0x4));
+
+	if (unitsToDeploy != 1)
+		return 0;
+
+	GET(TechnoClass* const, pThis, ESI);
+	auto const whatAmI = pThis->WhatAmI();
+
+	if (whatAmI == AbstractType::Infantry)
+	{
+		auto pInfantry = static_cast<InfantryClass*>(pThis);
+
+		if (!pInfantry->IsDeploying)
+		{
+			if (pInfantry->IsDeployed())
+				pThis->QueueVoice(pThis->GetTechnoType()->VoiceUndeploy);
+			else
+				pThis->QueueVoice(pThis->GetTechnoType()->VoiceDeploy);
+		}
+	}
+	else if (whatAmI == AbstractType::Unit)
+	{
+		auto pUnit = static_cast<UnitClass*>(pThis);
+
+		if (pUnit->TryToDeploy() || pUnit->Type->IsSimpleDeployer)
+			pThis->QueueVoice(pThis->GetTechnoType()->VoiceDeploy);
+	}
+
+	return 0;
+}
+
 #pragma endregion